Created framework for analysis of each frame
Idea is to have each frame analysed in a single pass with as little
recomputation and waste as possible. As such this is far more integrated
than previous sets of code.

Almost all new code is in the file analyse.cpp

Have implemented:
 - Find neighbours
 - Colour molecules
 - Neighbour histogram
 - contact histogram
 - Short order histogram

Partially Implemented:
 - MSD
 - Rotational relaxations

Have not yet got around to testing anything.