Working implementation of analyse
The code in the file analyse runs, producing results that appear to be
correct. I am yet to test the results against my previous
implementation.

This code runs somewhat slower, however it is checking every frame for
neighbours to generate time dependent plots. However there is a huge
reduction in the memory useage, especially on large problems.

Have Implemented:
 - Find neighbours
 - Colour molecules
 - Neighbour histogram
 - Contact histogram
 - Short Order Histogram
 - MSD
 - Diffusion Constant
 - Rotation relaxations
 - Rotation Relaxation time
 - Angles
 - Order
 - Short Order
 - Plot Frame

TODO:
 - Method to determine composition
 - Method to track movement of crystal front
 - Pairing
 - G(r)
 - Video