Using short range order for movie colouring
The short range ordering is used to determine whether the structure is
crystalline of liquid. Currently only looking at the antiparallel 1
structure (returns 4), this is the structure of the d = 1.637556
molecules.

Smoothing has also been added, molecules have to be liqiuid or solid for
two frames in a row to flip colour. This removes a lot of the noise from
the vibrational motion.