High and low mobility histograms, moved at max alpha
Creating high and low mobility contat and neighbour histograms to try
and find a link between the local structure and dynamics. Regions with
high contact numbers might be more mobile.

Moved plot is now at the max of alpha, where there should be the largest
inhomogenaety in the motion of particles