Adding new radial distribution and movie ouput
New ratio radial distribution, takes the ratio of the bond length to the
sum of the sigmas of the two intereating particles. This should result
in a peak at 1, with some type of distribution around it. This is to
check that by using a neighbour cutoff of 1.2 we are including all of
this distribution around 1.

Also editing movie output to output the velocity as the number of
contacts. The movie file can then be read into VMD and a small number of
molecules selected. The final visualisation could be the superposition
of all frames on top of each other.